Changes volume detach() method to use the deviceName
Previously, the detach() method assumed that the local deviceName
for an attached disk would match the volume name. While this is a
reasonable practice, equivalence is not required by GCE and there
are valid situations where having them be different makes sense.

This, in order to reliably detach a volume, we need the local deviceName
that was assigned when the volume was attached. This information is
only available in the instance metadata. See:
https://cloud.google.com/compute/docs/reference/rest/v1/instances/detachDisk